requirements.txt
Flask==2.3.3
pandas>=1.5.0,<2.1.0
python-dotenv==1.0.0
gunicorn==21.2.0
weasyprint==60.2
openpyxl==3.1.2
pytz==2023.3

